// resources/js/api/generated/models/meta.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Berth API
 * Berth: Opinionated docker compose stack management API
 * OpenAPI spec version: 1.0.0
 */

export interface Meta {
  /** @nullable */
  page?: number | null;
  /** @nullable */
  pageSize?: number | null;
  requestId?: string;
  /** @nullable */
  totalCount?: number | null;
}